extract ReportFetchController from ReportScreen
